Added Chunk + Item compatibility, corrected chunk generation and activation. Start of 2D Noise handling for heightmap

+++ Assets/_Project/Scripts/DataClasses/GameConfiguration.cs
@@ -6,6 +6,6 @@ namespace _Project.Scripts.DataClasses {
     [Serializable]
     public class GameConfiguration : ScriptableObject {
         [Range(1, 200)]
-        public float renderDistance;
+        public int renderDistance;
     }
 }
